requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
gunicorn
python-dotenv
psycopg2-binary
requests